fix(api): a cold start no longer refuses to diagnose restored shots
The cooldown read `diagnosed_at.get(shot_id, 0.0)`, so a shot this instance
had never answered was treated as answered at the clock's origin. That origin
is per sandbox: a freshly deployed revision reported `X-Dailies-Answer-Age: 44`
for a shot nobody had asked about, because nothing had stamped the map and the
age returned was `time.monotonic()` itself.

Every shot restored from the answer store therefore sat inside the five-minute
cooldown for the first five minutes of an instance's life. Cloud Run scales to
zero, so that is the normal path for a visitor arriving at an idle service: open
the board, press Diagnose, get told to wait for an answer nobody asked for.

Absence is now `None` and means "never answered here", which is what it always
meant.

The test needed the clock as an input to fail at all. A developer machine has
been up for days, so the same defect computes an age of several hundred thousand
seconds and sails past the cooldown; it reproduced only in the sandbox. Two
earlier drafts of the test passed against the broken code, the second because
`ShotStore.get` returns a deep copy and the restored diagnosis never reached the
store.

Also fixes four lint findings under ruff 0.16.4, one of which was a defect: two
tests shared a name in test_visual_qa.py, so the parser-level one was shadowed
and had never run. Renamed rather than deleted, since they cover different
layers. It passes.

+++ apps/api/dailies_api/duration_sample.py
@@ -75,7 +75,7 @@ def sample_from_buckets(buckets: Mapping[float, float]) -> list[float]:
             else:
                 representative = (previous_bound + bound) / 2.0
             if representative > 0:
-                sample.extend([representative] * int(round(in_bucket)))
+                sample.extend([representative] * round(in_bucket))
 
         if not math.isinf(bound):
             previous_bound = bound
